requirements.txt
pykalshi
pandas
pyarrow
requests
feedparser
python-dotenv
openai
google-generativeai
scikit-learn
numpy
